Replace the Live Preview heading when viewing a past recording
