+++ client/utils/formatters.js
@@ -1,0 +1,31 @@
+const fileSizeFormatter = new Intl.NumberFormat(undefined, {
+  maximumFractionDigits: 1,
+  minimumFractionDigits: 1,
+});
+
+const timestampFormatter = new Intl.DateTimeFormat(undefined, {
+  dateStyle: 'medium',
+  timeStyle: 'short',
+});
+
+export function formatFileSize(bytes) {
+  if (!Number.isFinite(bytes) || bytes < 0) return '';
+
+  const units = ['B', 'KB', 'MB', 'GB'];
+  let value = bytes;
+  let unitIndex = 0;
+
+  while (value >= 1024 && unitIndex < units.length - 1) {
+    value /= 1024;
+    unitIndex += 1;
+  }
+
+  return `${fileSizeFormatter.format(value)} ${units[unitIndex]}`;
+}
+
+export function formatTimestamp(value) {
+  if (!value) return null;
+
+  const date = new Date(value);
+  return Number.isNaN(date.getTime()) ? null : timestampFormatter.format(date);
+}
